third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/log.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] log float16 positive 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test log float16 actual 0 should be close enough to expected 4.15625 by ULP distance: expected a number less than or equal to 8 but got 17448
[FAIL] log float16 positive 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test log float16 actual 0 should be close enough to expected 4.15625 by ULP distance: expected a number less than or equal to 8 but got 17448
[FAIL] log float16 positive 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test log float16 actual 0 should be close enough to expected 4.15625 by ULP distance: expected a number less than or equal to 8 but got 17448
[FAIL] log float16 positive 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test log float16 actual 0 should be close enough to expected 4.15625 by ULP distance: expected a number less than or equal to 8 but got 17448
[FAIL] log float16 positive 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test log float16 actual 0 should be close enough to expected 4.15625 by ULP distance: expected a number less than or equal to 8 but got 17448
[FAIL] log float16 positive 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test log float16 actual 0 should be close enough to expected 4.15625 by ULP distance: expected a number less than or equal to 8 but got 17448
Harness: the test ran to completion.
